Fix for uncaught typeError (Issue 197)
The request library will return an undefined response object if a request response has a 0 length (see mikeal/request#414). This causes node-soap to return an uncaught typeError exception, immediately killing the node process, since it doesn't check if the response object is undefined before trying to retrieve the headers from it.