Taking ignoredNamespaces into account for objectToXML
If an object is converted into an xml representation (in wsdl.js, objectToXML) the element will never be prefixed with the namespace, even if one provided the corresponding option to createClient. The function only compares the given namespace with the constant property of wsdl.js.

This commit fixes that by comparing this.options.ignoredNamespaces, which is the overridden/concatenated version of this.ignoredNamespaces and the corresponding option, provided to createClient.