Add the incoming message object to the client 'response' event
If i really want to look at the actual response (and not only the body
that has already been stripped from everything but the actual SOAP
Envelope in lib/http.js), i need the `response` object, i.e. the
`IncomingMessage` object. Adding it to the 'response' event gives me the
opportunity to e.g. look at the headers and/or (in my case) the actual
body and start parsing a multipart response message
Update Readme and Test

Signed-off-by: Pattrick Hueper <phueper@hueper.net>