Support array of certs for ClientSSLSecurity ca.
In server configurations where there are problems with the server side
certificate chain, it is sometimes required that the client add multiple
certs into the trust repository. The lower-level https library supports
this transparently in the agentOptions but ClientSSLSecurity's type
detection of Buffer and String made it impossible to access directly.

This commit adds support for passing an array of certificates.