On error, call callback with error instead of message
As it stands, the callback for `WSDL` is sometimes called with a string. This error arises from the line `self.callback(e.message)`. This strips useful information (such as the stack trace) before sending the error back to the consumer, as well as violates JS best practices by throwing a string instead of an `Error`.

This PR corrects the issue by calling the callback with the error.