Add test for accepting array as parameter based on wsdl
v. 0.19.0 broke giving array as a parameter. It would now accept the "array"
with invalid syntax like this: {input: {PeriodList: {"PeriodType[]": {PeriodId: '1'}}}}
i.e. it wants the "[]" as the parameter name and it does not accept an array, but an
object. The invalid syntax then prints invalid XML like this:
"<PeriodList><PeriodType[]><PeriodId>1</PeriodId></PeriodType[]></PeriodList>"

This test shows old and correct syntax.

This test passes if you revert commit 5dbcb1d715843310c799419f260e0962722cd2fb
or disable if statement mentioned in
https://github.com/vpulim/node-soap/pull/914#pullrequestreview-40039598