Option return a soap fault instead of stack/error on bad request
In production, when a bad soap request is made, it should return a soap
fault and not a stack trace or text error message.

This change adds a WSDL option to return the fault, or continue to
return the stack trace or message.

A test case has been added to cover this new  functionality

also cleans up some jshint issues.