Avoid lodash where possible
Lodash is quote big project and it's required less with a time in modern
environments. In this diff I bumped node support to 8+ and used native
methods where possible instead of lodash. Later we can go further and
completely remove it from the project.

This can make the project much smaller.
https://packagephobia.com/result?p=soap