Change axios to be a peer dependency (#1179)
The usage of axios in this repository is very basic, but its versioning
is defined very tightly. This causes problems, because the interface
IOptions uses AxiosInstance as its property, preventing the clients of
this library from updating their axios version beyond the limits set in
this library. Typescript compiler gives type error when giving higher
version AxiosInstance for IOptions.

By setting axios as a peer dependency, we would allow the clients of
this library to install other versions of axios if they so choose.