Removed double await in async and fixed response (#1185)
* Removed double await in async 

There's no need to call await again on the result object since result is already resolved.

* Returning just the JSON response in await example 

In the example for async, logging result returns an array containing not just the JSON response but also the rawResponse, rawRequest e.t.c. 
This can be confusing for users who try the async example after trying the non-async examples which only return the JSON response. 

I added a subscript to the 0th index so that that's what's logged.