Speed up parsing with many namespaces (#1322)
The XML schema instance handling of the parsing code behaves linearly with regards to
the number of namespaces. If there are many namespaces in the response (in our
case it was over 200, almost all of them unused) this causes large
slowdowns, which are especially problematic when parsing large amounts
of XML in resource constrained environments.

The changes sped up parsing up by more than 5x in our case.